Fixed application of flat flied

+++ FlatFieldCalibration/FunctionLibrary/Apply_Flat_Field_Calibration.m
@@ -1,4 +1,3 @@
 function calibrated = Apply_Flat_Field_Calibration(uncalibrated, gain, offset)
 
-
-calibrated = (uncalibrated-offset)./gain;
+calibrated = (double(uncalibrated)-offset)./gain;
